+++ webapp/src/components/posts/_index.ts
@@ -1,2 +1,4 @@
 export { VideoView } from "./video";
 export type { IVideoViewProps } from "./video";
+export { NewVideoForm } from "./new-video";
+export type { INewVideoFormProps } from "./new-video";
